simple-multi-tenancy/pyproject.toml
[project]
name = "simple-multi-tenancy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.128.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
]
